refactor: comment 관련 컴포넌트 comment entities 로 슬라이스 분리
